+++ packages/web3-core/src/web3_request_manager.ts
@@ -87,14 +87,25 @@ export class Web3RequestManager<
 		this.useRpcCallSpecification = useRpcCallSpecification;
 	}
 
+	/**
+	 * Will return all available providers
+	 */
 	public static get providers() {
 		return availableProviders;
 	}
 
+	/**
+	 * Will return the current provider.
+	 *
+	 * @returns Returns the current provider
+	 */
 	public get provider() {
 		return this._provider;
 	}
 
+	/**
+	 * Will return all available providers
+	 */
 	// eslint-disable-next-line class-methods-use-this
 	public get providers() {
 		return availableProviders;
@@ -103,6 +114,8 @@ export class Web3RequestManager<
 	/**
 	 * Use to set provider. Provider can be a provider instance or a string.
 	 * To set IPC provider as a string please use the IPC socket file which name ends with .ipc
+	 *
+	 * @param provider - The provider to set
 	 */
 	public setProvider(provider?: SupportedProviders<API> | string, net?: Socket): boolean {
 		let newProvider: SupportedProviders<API> | undefined;
@@ -138,6 +151,15 @@ export class Web3RequestManager<
 		return true;
 	}
 
+	/**
+	 *
+	 * Will execute a request
+	 *
+	 * @param request - {@link Web3APIRequest} The request to send
+	 *
+	 * @returns The response of the request {@link ResponseType}. If there is error
+	 * in the response, will throw an error
+	 */
 	public async send<
 		Method extends Web3APIMethod<API>,
 		ResponseType = Web3APIReturnType<API, Method>,
@@ -151,6 +173,11 @@ export class Web3RequestManager<
 		throw new ResponseError(response);
 	}
 
+	/**
+	 * Same as send, but, will execute a batch of requests
+	 *
+	 * @param request {@link JsonRpcBatchRequest} The batch request to send
+	 */
 	public async sendBatch(request: JsonRpcBatchRequest): Promise<JsonRpcBatchResponse<unknown>> {
 		const response = await this._sendRequest<never, never>(request);