+++ packages/web3-eth/test/unit/rpc_method_wrappers/fixtures/get_fee_history.ts
@@ -65,6 +65,14 @@ export const testData: TestData[] = [
 		'blockCount = "0x4b7", newestBlock = BlockTags.PENDING, rewardPercentiles = ["0x0"]',
 		['0x4b7', BlockTags.PENDING, ['0x0']],
 	],
+	[
+		'blockCount = "0x4b7", newestBlock = BlockTags.SAFE, rewardPercentiles = ["0x0"]',
+		['0x4b7', BlockTags.SAFE, ['0x0']],
+	],
+	[
+		'blockCount = "0x4b7", newestBlock = BlockTags.FINALIZED, rewardPercentiles = ["0x0"]',
+		['0x4b7', BlockTags.FINALIZED, ['0x0']],
+	],
 	// blockNumber = Numbers
 	[
 		'blockCount = "0x4b7", newestBlock = 1207, rewardPercentiles = ["0x0"]',