+++ packages/web3/test/unit/version.test.ts
@@ -15,18 +15,11 @@ You should have received a copy of the GNU Lesser General Public License
 along with web3.js.  If not, see <http://www.gnu.org/licenses/>.
 */
 
-import { BlockTags } from 'web3-types';
-import { isUInt } from './numbers';
+import { Web3PkgInfo } from '../../src/version';
+import packageFile from '../../package.json';
 
-export const isBlockNumber = (value: string | number | bigint): boolean => isUInt(value);
-
-/**
- * Returns true if the given blockNumber is 'latest', 'pending', 'earliest, 'safe' or 'finalized'
- */
-export const isBlockTag = (value: string) => Object.values(BlockTags).includes(value as BlockTags);
-
-/**
- * Returns true if given value is valid hex string and not negative, or is a valid BlockTag
- */
-export const isBlockNumberOrTag = (value: string | number | bigint) =>
-	isBlockTag(value as string) || isBlockNumber(value);
+describe('web3 package info', () => {
+	it('should Web3PkgInfo.version returns the same version set at package.json', () => {
+		expect(packageFile.version).toEqual(Web3PkgInfo.version);
+	});
+});